Create gizi_Image_Classificaiton v1
Reducing the categories from our old dataset from 10 to 2. The  result of this binary classification is satisfying enough. The model can classify if the image contain apple or orange. But the model is better performing on the group of apple or orange rather than single apple or orange. The model is a CNN model built from scratch.

Co-Authored-By: gabrielleaudrey <30653750+gabrielleaudrey@users.noreply.github.com>