Create gizi_Image_Classification v2
We tried transfer learning method on the dataset with 2 categories of apple or orange. We use inception v3 as the base model for feature extractor. The result is this model performing with higher accuracy and much faster in training. Still better performing in group of apples or oranges rather than single apple or orange.

Co-Authored-By: gabrielleaudrey <30653750+gabrielleaudrey@users.noreply.github.com>